Initial commit.
This project is a set of examples of FFI between Rust and C.
It uses CMake to build the C, and CMake to call Cargo to build the Rust.
This is not intended to be a Rust, C, or CMake tutorial, merely an example of using them together. That said, comments should explain to C programmers what the Rust code is doing.

This has only been tested on Linux.